REQ[#46401]: 登录越权告警
[Description]: 修改接口权限。审计服务接口添加权限。录像文件解码后websocket推给前端。
[Test Suggestion]:无

+++ iot-center-platform/src/main/java/com/yjh/platform/module/user/service/TRecordFileInfoService.java
@@ -3,6 +3,7 @@ package com.yjh.platform.module.user.service;
 import com.yjh.platform.module.user.dao.TRecordFileInfoDao;
 import com.yjh.platform.module.user.entity.TRecordFileInfo;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,11 +25,13 @@ public class TRecordFileInfoService {
     @Transactional(rollbackFor = Exception.class)
     public int deleteByPrimaryId(Long id) {
         TRecordFileInfo tRecordFileInfo = tRecordFileInfoDao.selectByPrimaryId(id);
-        String absolutePathUrl = "rm -rf " + tRecordFileInfo.getAbsoluteFilePath();
-        try {
-            Runtime.getRuntime().exec(absolutePathUrl);
-        }catch (Exception e){
-            log.info("录像文件删除失败", e);
+        if (StringUtils.isNotEmpty(tRecordFileInfo.getAbsoluteFilePath())){
+            String absolutePathUrl = "rm -rf " + tRecordFileInfo.getAbsoluteFilePath();
+            try {
+                Runtime.getRuntime().exec(absolutePathUrl);
+            }catch (Exception e){
+                log.info("录像文件删除失败", e);
+            }
         }
         return tRecordFileInfoDao.deleteByPrimaryId(id);
     }
